refactor(models): SubmissionStatus + SubmissionResult belong with H1
The status enum is H1's report-state taxonomy (triaged / duplicate /
informative / not_applicable are H1 vocabulary, not generic submission
states); SubmissionResult carries h1_url and is only ever produced by
tools/h1_api.submit_report. Both belong with the H1 surface in models/h1.py.

Three importers migrate to "from models.h1 import ..."; no back-compat
re-export from the package root. DisclosureReport stays in models/__init__
for now - it has H1 coupling (weakness_id, structured_scope_id,
programme_handle) but its body_markdown / impact_statement are generic
authored content; deferring that call until there's a second platform or
an explicit ask.
